refactor for testability
